+++ app/src/main/java/com/dewdrop623/androidcrypt/JNICallbackInterface.java
@@ -1,6 +1,6 @@
 package com.dewdrop623.androidcrypt;
 
 public abstract class JNICallbackInterface {
-    public abstract void progressCallback(int totalBytes);
+    public abstract void progressCallback(long totalBytes);
     public abstract void completedCallback(int status);
 }
